+++ lib/tab_navigation_item.dart
@@ -1,5 +1,5 @@
 import 'package:bikeangletest/pages/logbook/logbook.dart';
-import 'package:bikeangletest/pages/recording.dart';
+import 'package:bikeangletest/pages/recording/recording.dart';
 import 'package:flutter/material.dart';
 
 class TabNavigationItem {
